pyproject.toml
[tool.poetry]
name = "resin-api"
version = "0.1.0"
description = ""
authors = ["Maud Royer <hello@maudroyer.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
pillow = "^10.3.0"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"

[tool.poetry.dev-dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
